Finish the entire recipes project
Got distracted while working and only remembered to commit at the end.

Here is a summarized description of everything.

Modify README.md to describe the project.
Add an image folder and add recipe images to it.
Create three pages for three different recipes.
Fill recipe pages with content.
Add links to the recipe pages in the index page.